% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_r_binding.R
\name{serialize_gaussian_kernel}
\alias{serialize_gaussian_kernel}
\title{Serialize GaussianKernel.}
\usage{
serialize_gaussian_kernel(filename, model_in = NULL)
}
\arguments{
\item{filename}{Input filename.}

\item{model_in}{Input model.}
}
\description{
Serialize a model to the given filename.
}
