% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_r_binding.R
\name{serialize_gaussian_kernel_to_xml}
\alias{serialize_gaussian_kernel_to_xml}
\title{Serialize GaussianKernel to xml}
\usage{
serialize_gaussian_kernel_to_xml(model_in = NULL)
}
\arguments{
\item{model_in}{Input model.}
}
\value{
transformed_model
}
\description{
Extract serialized information for model.
}
