% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lars.R
\name{serialize_lars}
\alias{serialize_lars}
\title{Serialize LARS.}
\usage{
serialize_lars(filename, model_in = NULL)
}
\arguments{
\item{filename}{Input filename.}

\item{model_in}{Input model.}
}
\description{
Serialize a model to the given filename.
}
