% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lars.R
\name{serialize_lars_to_xml}
\alias{serialize_lars_to_xml}
\title{Serialize LARS to xml}
\usage{
serialize_lars_to_xml(model_in = NULL)
}
\arguments{
\item{model_in}{Input model.}
}
\value{
transformed_model
}
\description{
Extract serialized information for model.
}
