% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_model}
\alias{transform_model}
\title{Transform Model}
\usage{
transform_model(model)
}
\arguments{
\item{model}{Input model xml.}
}
\value{
transformed_model
}
\description{
Extract serialized information for model.
}
