% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_r_binding.R
\name{unserialize_gaussian_kernel}
\alias{unserialize_gaussian_kernel}
\title{Unserialize GaussianKernel.}
\usage{
unserialize_gaussian_kernel(filename)
}
\arguments{
\item{filename}{Input filename.}
}
\value{
model_ptr Output model.
}
\description{
Unserialize a model to the given filename.
}
