% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lars.R
\name{unserialize_lars}
\alias{unserialize_lars}
\title{Unserialize LARS.}
\usage{
unserialize_lars(filename)
}
\arguments{
\item{filename}{Input filename.}
}
\value{
model_ptr Output model.
}
\description{
Unserialize a model to the given filename.
}
