% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_r_binding.R
\name{test_r_binding}
\alias{test_r_binding}
\title{Run to test R binding functionality}
\usage{
test_r_binding(
  double_in,
  int_in,
  string_in,
  matrix_in = matrix(NA),
  build_model = FALSE,
  flag1 = FALSE,
  flag2 = FALSE,
  model_in = NULL,
  copy_all_inputs = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{double_in}{Input double, must be 4.0.}

\item{int_in}{Input int, must be 12.}

\item{string_in}{Input string, must be 'hello'.}

\item{matrix_in}{Input matrix.}

\item{build_model}{If true, a model will be returned.}

\item{flag1}{Input flag, must be specified.}

\item{flag2}{Input flag, must not be specified.}

\item{model_in}{Input model.}

\item{copy_all_inputs}{If specified, all input parameters will be
deep copied before the method is run.  This is useful for debugging
problems where the input parameters are being modified by the algorithm,
but can slow down the code.}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value `FALSE`.}
}
\value{
A list with several components:
\item{double_out}{ Output double, will be 5.0.}
\item{int_out}{ Output int, will be 13.}
\item{string_out}{ Output string, will be 'hello2'.}
\item{matrix_out}{ Output matrix.}
\item{model_out}{ Output model, with twice the bandwidth.}
\item{model_bw_out}{ The bandwidth of the model.}
}
\description{
A simple program to test R binding functionality.
}
\examples{
## test_r_binding:
x <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15), nrow = 5)
out1 <- test_r_binding(4.0, 12, "hello", flag1 = TRUE, build_model = TRUE)
out2 <- test_r_binding(4.0, 12, "hello", x, TRUE, TRUE, model_in = out1$model_out)
}
