% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_r_binding.R
\name{testRBinding}
\alias{testRBinding}
\title{Run to test R binding functionality}
\usage{
testRBinding(
  doubleIn,
  intIn,
  stringIn,
  matrixIn = matrix(NA),
  buildModel = FALSE,
  flag1 = FALSE,
  flag2 = FALSE,
  modelIn = NULL
)
}
\arguments{
\item{doubleIn}{Input double, must be 4.0.}

\item{intIn}{Input int, must be 12.}

\item{stringIn}{Input string, must be 'hello'.}

\item{matrixIn}{Input matrix.}

\item{buildModel}{If true, a model will be returned.}

\item{flag1}{Input flag, must be specified.}

\item{flag2}{Input flag, must not be specified.}

\item{modelIn}{Input model.}
}
\value{
A list with several components: Output double, will be 5.0,
Output int, will be 13, Output string, will be 'hello2', Output matrix, 
Output model, with twice the bandwidth, The bandwidth of the model.
}
\description{
A simple program to test R binding functionality.
}
\examples{
## testRBinding:
x <- matrix ( c( 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15), nrow = 5)
out1 <- testRBinding(4.0, 12, 'hello', flag1 = TRUE, buildModel = TRUE)
out2 <- testRBinding(4.0, 12 , 'hello', x, TRUE, TRUE, modelIn = out1$modelOut)
}
